maj doli16
Former-commit-id: 80e9cb1dc608b3a6882c43ccb8a3284c425f204f

+++ htdocs/bimpcore/classes/BimpDb.php
@@ -30,6 +30,8 @@ class BimpDb
             $fields .= $name;
             if (is_int($value)) {
                 $values .= (int) $value;
+            } elseif (is_null($value)) {
+                $values .= 'NULL';
             } else {
                 if (is_array($value)) {
                     $value = json_encode($value);
@@ -42,9 +44,9 @@ class BimpDb
 
         $sql = 'INSERT INTO ' . MAIN_DB_PREFIX . $table . $fields . $values;
 
-        if (BimpDebug::isActive('bimpcore/objects/print_insert_sql')) {
-            echo 'SQL: ' . $sql . '<br/>';
-        }
+//        if (BimpDebug::isActive()) {
+//            BimpDebug::addDebug('bimpdb_sql', 'INSERT - ' . $table, BimpRender::renderSql($sql));
+//        }
         $result = $this->db->query($sql);
         if ($result > 0) {
             if ($return_id) {
@@ -53,7 +55,7 @@ class BimpDb
             return 1;
         }
 
-        $this->logSqlError();
+        $this->logSqlError($sql);
         return 0;
     }
 
@@ -80,39 +82,32 @@ class BimpDb
             if (is_array($value)) {
                 $value = json_encode($value);
             }
-            $sql .= '"' . $this->db->escape($value) . '"';
+
+            if (is_null($value)) {
+                $sql .= 'NULL';
+            } else {
+                $sql .= '"' . $this->db->escape($value) . '"';
+            }
         }
         $sql .= ' WHERE ' . $where;
 
-        if (BimpDebug::isActive('bimpcore/objects/print_update_sql')) {
-            echo 'SQL: ' . $sql . '<br/>';
-        }
+//        if (BimpDebug::isActive()) {
+//            BimpDebug::addDebug('bimpdb_sql', 'UPDATE - ' . $table, BimpRender::renderSql($sql));
+//        }
 
         return $this->execute($sql);
     }
 
     public function execute($sql)
     {
-        $transac = (stripos(trim($sql), "SELECT") === 0) ? 0 : 1;
-        if ($transac)
-            $this->db->begin();
-
         $result = $this->db->query($sql);
-
-        if ($transac) {
-            if ($result > 0) {
-                $this->db->commit();
-            } else {
-                $this->db->rollback();
-            }
-        }
         if (!$result)
-            $this->logSqlError();
+            $this->logSqlError($sql);
 
         return $result;
     }
 
-    public function executeS($sql, $return = 'object')
+    public function executeS($sql, $return = 'object', $returned_field = null)
     {
         $result = $this->db->query($sql);
 
@@ -126,12 +121,17 @@ class BimpDb
                         break;
 
                     case 'array':
-                        $rows[] = $this->db->fetch_array($result);
+                        if (is_null($returned_field)) {
+                            $rows[] = $this->db->fetch_array($result);
+                        } else {
+                            $res = $this->db->fetch_array($result);
+                            $rows[] = $res[$returned_field];
+                        }
                         break;
                 }
             }
         } else {
-            $this->logSqlError();
+            $this->logSqlError($sql);
         }
 
         $this->db->free($result);
@@ -139,26 +139,36 @@ class BimpDb
         return $rows;
     }
 
-    public function executeFile($file)
+    public function executeFile($file, &$errors = array())
     {
         if (file_exists($file)) {
             $sql = file_get_contents($file);
             $sql = str_replace("llx_", MAIN_DB_PREFIX, $sql);
             $sql = str_replace("MAIN_DB_PREFIX", MAIN_DB_PREFIX, $sql);
             if ($sql) {
-                $tabSql = explode(";", $sql);
+//                $sql = str_replace("; \n", ";\n", $sql);
+                $sql = preg_replace("/;( )*\n/U", ";\n", $sql);
+                $tabSql = explode(";\n", $sql);
                 foreach ($tabSql as $req) {
                     if ($req != "")
-                        if ($result = $this->execute($req) < 0)
+                        if ($result = $this->execute($req) < 0) {
+                            BimpCore::addlog('Erreur SQL maj', 3, 'sql', null, array(
+                                'Requête' => (!is_null($req) ? $req : ''),
+                                'Erreur'  => $this->err()
+                            ));
+                            $errors[] = 'Echec requête "' . $req . '" - ' . $this->err();
                             return false;
+                        }
                 }
             }
             return true;
         }
+
+        $errors[] = 'Le fichier "' . $file . '" n\'existe pas';
         return false;
     }
 
-    public function getRows($table, $where = '1', $limit = null, $return = 'object', $fields = null, $order_by = null, $order_way = null)
+    public function getRows($table, $where = '1', $limit = null, $return = 'object', $fields = null, $order_by = null, $order_way = null, $joins = array())
     {
         $sql = 'SELECT ';
 
@@ -170,19 +180,35 @@ class BimpDb
                 } else {
                     $fl = false;
                 }
-                $sql .= '`' . $field . '`';
+
+                if (!preg_match('/[\. ]/', $field)) {
+                    $sql .= '`' . $field . '`';
+                } else {
+                    $sql .= $field;
+                }
             }
         } else {
             $sql .= '*';
         }
 
         $sql .= ' FROM ' . MAIN_DB_PREFIX . $table;
+
+        foreach ($joins as $join) {
+            $sql .= ' LEFT JOIN ' . MAIN_DB_PREFIX . $join['table'] . ' ' . $join['alias'];
+            $sql .= ' ON ' . $join['on'];
+        }
+
         $sql .= ' WHERE ' . $where;
 
         if (!is_null($order_by)) {
-            $sql .= ' ORDER BY `' . $order_by . '`';
+            $sql .= ' ORDER BY ';
+            if (!preg_match('/\./', $order_by)) {
+                $sql .= '`' . $order_by . '`';
+            } else {
+                $sql .= $order_by;
+            }
             if (!is_null($order_way)) {
-                $sql .= strtoupper($order_way);
+                $sql .= ' ' . strtoupper($order_way);
             }
         }
 
@@ -193,7 +219,7 @@ class BimpDb
         return $this->executeS($sql, $return);
     }
 
-    public function getRow($table, $where = '1', $fields = null, $return = 'object')
+    public function getRow($table, $where = '1', $fields = null, $return = 'object', $order_by = '', $order_way = 'ASC')
     {
         $sql = 'SELECT ';
 
@@ -212,7 +238,20 @@ class BimpDb
         }
 
         $sql .= ' FROM ' . MAIN_DB_PREFIX . $table;
-        $sql .= ' WHERE ' . $where . ' LIMIT 1';
+        $sql .= ' WHERE ' . $where;
+
+        if ($order_by) {
+            $sql .= ' ORDER BY `' . $order_by . '` ' . $order_way;
+        }
+
+        $sql .= ' LIMIT 1';
+
+//        if ($table == 'stock_mouvement') {
+//            die($sql);
+//        }
+
+
+
         $result = $this->db->query($sql);
         if ($result && $this->db->num_rows($result)) {
             if ($return === 'object') {
@@ -243,9 +282,13 @@ class BimpDb
         return null;
     }
 
-    public function getValue($table, $field, $where = '1')
+    public function getValue($table, $field, $where = '1', $order_by = '', $order_way = 'DESC')
     {
-        $sql = 'SELECT `' . $field . '` FROM ' . MAIN_DB_PREFIX . $table . ' WHERE ' . $where . ' LIMIT 1';
+        $sql = 'SELECT `' . $field . '` FROM ' . MAIN_DB_PREFIX . $table . ' WHERE ' . $where;
+        if ($order_by) {
+            $sql .= ' ORDER BY `' . $order_by . '` ' . $order_way;
+        }
+        $sql .= ' LIMIT 1';
         $result = $this->db->query($sql);
 
         if ($result && $this->db->num_rows($result)) {
@@ -298,6 +341,19 @@ class BimpDb
         return 0;
     }
 
+    public function getAvg($table, $field, $where = '1')
+    {
+        $sql = 'SELECT AVG(`' . $field . '`) as avg FROM ' . MAIN_DB_PREFIX . $table . ' WHERE ' . $where;
+
+        $result = $this->executeS($sql, 'array');
+
+        if (isset($result[0]['avg'])) {
+            return $result[0]['avg'];
+        }
+
+        return null;
+    }
+
     public function getValues($table, $field, $where, $limit = null)
     {
         $sql = 'SELECT `' . $field . '` FROM ' . MAIN_DB_PREFIX . $table . ' WHERE ' . $where;
@@ -320,8 +376,20 @@ class BimpDb
         return $rows;
     }
 
+    public function rowExists($table, $id, $primary = 'id')
+    {
+        return ((int) $this->getCount($table, $primary . ' = ' . $id, $primary) > 0 ? true : false);
+    }
+
     public function delete($table, $where)
     {
+        if (!(string) $where || (string) $where == '1') {
+            BimpCore::addlog('Delete SQL sans WHERE', Bimp_Log::BIMP_LOG_URGENT, 'bimpcore', null, array(
+                'table' => $table
+            ));
+            return 0;
+        }
+
         $sql = 'DELETE FROM ' . MAIN_DB_PREFIX . $table;
         $sql .= ' WHERE ' . $where;
 
@@ -330,11 +398,14 @@ class BimpDb
 
     protected function logSqlError($sql = null)
     {
-        $msg = 'Erreur SQL' . "\n";
-        if (!is_null($sql)) {
-            $msg .= 'Requête: ' . $sql . "\n";
-        }
-        $msg .= 'Msg SQL: ' . $this->db->lasterror();
-        dol_syslog($msg, 3);
+//        BimpCore::addlog('Erreur SQL', 3, 'sql', null, array(
+//            'Requête' => (!is_null($sql) ? $sql : ''),
+//            'Erreur'  => $this->db->lasterror()
+//        ));
+    }
+
+    public function err()
+    {
+        return $this->db->lasterror();
     }
 }
